# Optimize numeric-to-string coercion in StringArrayDecoder