--- a/arrow-json/src/reader/string_array.rs
+++ b/arrow-json/src/reader/string_array.rs
@@ -109,7 +109,8 @@ impl<O: OffsetSizeTrait> ArrayDecoder for StringArrayDecoder<O> {
                     builder.append_value(FALSE);
                 }
                 TapeElement::Number(idx) if coerce_primitive => {
-                    builder.append_value(tape.get_string(idx));
+                    let s = self.write_number(tape.get_string(idx));
+                    builder.append_value(s);
                 }
                 TapeElement::I64(high) if coerce_primitive => match tape.get(p + 1) {
                     TapeElement::I32(low) => {