@ndemir 2025-03-12
I see that the performance is consistently BETTER even though it is not hitting the refactored code.
Will investigate.